+++ modules/core.lua
@@ -1,4 +1,4 @@
 -- Datos basicos que usa todo el addon.
 EZOMetter.ADDON_NAME    = "EZOMetter"
-EZOMetter.ADDON_VERSION = "0.1.32"
+EZOMetter.ADDON_VERSION = "0.1.33"
 EZOMetter.AUTHOR        = "@Zuriplayer"